change filenames in gitops
